Move complementary functions from index to a new utils file
