+++ src/index.ts
@@ -24,44 +24,11 @@ entonces debe ir antes.
 Nota: todos los números de la lista son positivos y la lista puede estar vacia.
 
 */
+import { getSortedListByWeight } from './utils'
 
 const sortListByWeight = (input: string): string => {
     const array: string[] = input.split(' ')
     return getSortedListByWeight(array)
 }
 
-const sumOfDigits = (numberA: number, numberB: number): number => {
-    return numberA + numberB
-}
-
-const weight = (number: string): number => {
-    return number
-            .split('')
-            .map(x => Number(x))
-            .reduce(sumOfDigits)
-}
-
-const compareWeight = (stringA: string, stringB: string): number => {
-    const weightA: number = weight(stringA)
-    const weightB: number = weight(stringB)
-    const digitsHaveSameWeight: boolean = weightA === weightB
-
-    if (digitsHaveSameWeight) {
-        return stringA.localeCompare(stringB)
-    }
-    return weightA - weightB
-}
-
-const getSortedListByWeight = (array: string[]): string => {
-    return array
-            .sort(compareWeight)
-            .join(' ')
-}
-
-export {
-    sortListByWeight,
-    sumOfDigits,
-    compareWeight,
-    weight,
-    getSortedListByWeight
-}
+export default sortListByWeight
